engine_traits: move KvEngines to its own module
