+++ components/engine_traits/src/lib.rs
@@ -22,7 +22,7 @@ pub use crate::engine::*;
 mod options;
 pub use crate::options::*;
 mod db_options;
-mod util;
+pub mod util;
 pub use crate::db_options::*;
 mod cf_handle;
 pub use crate::cf_handle::*;
@@ -30,6 +30,8 @@ mod cf_options;
 pub use crate::cf_options::*;
 mod import;
 pub use import::*;
+mod engines;
+pub use engines::*;
 
 pub const DATA_KEY_PREFIX_LEN: usize = 1;
 
@@ -37,43 +39,3 @@ pub const DATA_KEY_PREFIX_LEN: usize = 1;
 // reduce OLTP QPS by 30% ~ 60%. We found that 32K is a proper choice.
 pub const MAX_DELETE_BATCH_SIZE: usize = 32 * 1024;
 
-#[derive(Clone, Debug)]
-pub struct KvEngines<K, R> {
-    pub kv: K,
-    pub raft: R,
-    pub shared_block_cache: bool,
-}
-
-impl<K: KvEngine, R: KvEngine> KvEngines<K, R> {
-    pub fn new(kv_engine: K, raft_engine: R, shared_block_cache: bool) -> Self {
-        KvEngines {
-            kv: kv_engine,
-            raft: raft_engine,
-            shared_block_cache,
-        }
-    }
-
-    pub fn write_kv(&self, wb: &K::Batch) -> Result<()> {
-        self.kv.write(wb)
-    }
-
-    pub fn write_kv_opt(&self, wb: &K::Batch, opts: &WriteOptions) -> Result<()> {
-        self.kv.write_opt(opts, wb)
-    }
-
-    pub fn sync_kv(&self) -> Result<()> {
-        self.kv.sync()
-    }
-
-    pub fn write_raft(&self, wb: &R::Batch) -> Result<()> {
-        self.raft.write(wb)
-    }
-
-    pub fn write_raft_opt(&self, wb: &R::Batch, opts: &WriteOptions) -> Result<()> {
-        self.raft.write_opt(opts, wb)
-    }
-
-    pub fn sync_raft(&self) -> Result<()> {
-        self.raft.sync()
-    }
-}
